services/resume_jd_matcher/requirements.txt
fastapi
uvicorn
PyMuPDF
python-dotenv
google-generativeai
python-multipart
